MNT: rename DispersionModel.py to dispersion_model.py
